UserChat component seperated
